breakout fetch example into a seperate component.
