+++ README.md
@@ -30,3 +30,7 @@ them, take them out of package.json or don't use them.
 
 - AJAX: Github Fetch https://github.com/github/fetch
 - Router: React-Router https://github.com/reactjs/react-router
+
+# Compatibility
+sorry, fetch doesn't work with IE11 without a promise  polyfill. Since I'm trying
+to keep everything to a minimum, I'm not including it
